# A few perf improvments for Tracing