--- a/tracing/src/main/java/com/palantir/remoting3/tracing/Tracers.java
+++ b/tracing/src/main/java/com/palantir/remoting3/tracing/Tracers.java
@@ -16,7 +16,6 @@
 
 package com.palantir.remoting3.tracing;
 
-import com.google.common.base.Strings;
 import java.util.Optional;
 import java.util.concurrent.Callable;
 import java.util.concurrent.ExecutorService;
@@ -27,6 +26,8 @@ import java.util.concurrent.ThreadLocalRandom;
 public final class Tracers {
     /** The key under which trace ids are inserted into SLF4J {@link org.slf4j.MDC MDCs}. */
     public static final String TRACE_ID_KEY = "traceId";
+    private static final char[] HEX_DIGITS =
+            {'0', '1', '2', '3', '4', '5', '6', '7', '8', '9', 'a', 'b', 'c', 'd', 'e', 'f'};
 
     private Tracers() {}
 
@@ -36,7 +37,24 @@ public final class Tracers {
     }
 
     static String longToPaddedHex(long number) {
-        return Strings.padStart(Long.toHexString(number), 16, '0');
+        char[] data = new char[16];
+        data[0] = HEX_DIGITS[(int) ((number >> 60) & 0xF)];
+        data[1] = HEX_DIGITS[(int) ((number >> 56) & 0xF)];
+        data[2] = HEX_DIGITS[(int) ((number >> 52) & 0xF)];
+        data[3] = HEX_DIGITS[(int) ((number >> 48) & 0xF)];
+        data[4] = HEX_DIGITS[(int) ((number >> 44) & 0xF)];
+        data[5] = HEX_DIGITS[(int) ((number >> 40) & 0xF)];
+        data[6] = HEX_DIGITS[(int) ((number >> 36) & 0xF)];
+        data[7] = HEX_DIGITS[(int) ((number >> 32) & 0xF)];
+        data[8] = HEX_DIGITS[(int) ((number >> 28) & 0xF)];
+        data[9] = HEX_DIGITS[(int) ((number >> 24) & 0xF)];
+        data[10] = HEX_DIGITS[(int) ((number >> 20) & 0xF)];
+        data[11] = HEX_DIGITS[(int) ((number >> 16) & 0xF)];
+        data[12] = HEX_DIGITS[(int) ((number >> 12) & 0xF)];
+        data[13] = HEX_DIGITS[(int) ((number >> 8) & 0xF)];
+        data[14] = HEX_DIGITS[(int) ((number >> 4) & 0xF)];
+        data[15] = HEX_DIGITS[(int) ((number >> 0) & 0xF)];
+        return new String(data);
     }
 
     /**