I want Tritium to trace by default, but the overhead seemed a little excessive (700ns per method invocation).

Here, we have 3 improvements:

First, don't iterate over a hashmap.

Second, handcoded 'toString' method.

Third, 'completeTraceFast' which avoids a bunch of computation if I'm
not going to use it (e.g. a System.nanoTime call which is 100 cycles
plus).

For changes (with the Tritium benchmark):

Unchanged: 711ns
With iterating over the list not the hashmap: 666ns
With the custom toHexString: 588ns

With the shortcircuit and a 1% sampling collector:

After all changes but shortcircuit: 256ns
With shortcircuit only: 220ns
After all changes: 198ns