@j-baker 2017-10-24
Btw, those overheads include the cost of the dynamic proxy - the additional overhead of the tracing are now low as opposed to high.